add: separate Auth file
